requirements.txt
aiohttp==1.0.5
async-timeout==2.0.1
chardet==3.0.4
discord.py==0.16.12
multidict==4.3.1
websockets==3.4
